requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# py_http_api_test/jmespath_custom.py: 7,8
jmespath == 0.9.3

# py_http_api_test/http_test.py: 9
nose == 1.3.7

# py_http_api_test/tools.py: 11
# tests/tools_test.py: 6,7
pycrypto == 2.6.1

# py_http_api_test/http_session.py: 4,7,8,9
requests == 2.18.4

# py_http_api_test/http_test.py: 10
ruamel.yaml == 0.15.34
